Make sure conditions are displayed if applicable
Created getConditions(application) function to use across multiple
routes.

Fixed details view display logic to check for conditions object
instead of status.offered.conditions.

Normalised stub data so that standard and non standard conditions
are consistently defined across different applications.
