+++ app/views/application/index.njk
@@ -168,7 +168,7 @@
     }]
   }) }}
 
-  {% if application.status.offered.conditions or application.status.offered["standard-conditions"] %}
+  {% if conditions %}
     <div class="govuk-grid-row">
       <div class="govuk-grid-column-two-thirds">
         <h2 class="govuk-heading-m govuk-!-font-size-27 govuk-!-margin-top-8">Conditions</h2>
